+++ examples/basic/slider_browser.py
@@ -1,6 +1,6 @@
 """Mouse hind limb growth from day 10 9h to day 15 21h"""
 from vedo import settings, dataurl, load
-from vedo import Text2D, Plotter, Picture, Axes, Line
+from vedo import Text2D, Plotter, Image, Axes, Line
 
 
 def sliderfunc(widget, event):
@@ -18,7 +18,7 @@ settings.default_font = "Glasgo"
 
 plt = Plotter(bg="blackboard")
 plt += Text2D(__doc__, pos="top-center", s=1.2, c="w")
-plt += Picture(dataurl + "images/limbs_tc.jpg").scale(0.0154).y(10)
+plt += Image(dataurl + "images/limbs_tc.jpg").scale(0.0154).y(10)
 plt += Line([(0, 8), (0, 10), (28.6, 10), (4.5, 8)], c="gray")
 plt += Axes(objs[-1])
 plt += objs[0]