+++ examples/volumetric/image_rgba.py
@@ -1,7 +1,7 @@
 """Example plot of 2 images containing an
 alpha channel for modulating the opacity"""
 #Credits: https://github.com/ilorevilo
-from vedo import Picture, show
+from vedo import Image, show
 import numpy as np
 
 rgbaimage1 = np.random.rand(50, 50, 4) * 255
@@ -10,9 +10,9 @@ rgbaimage1[:, :, 3] = alpharamp
 rgbaimage2 = np.random.rand(50, 50, 4) * 255
 rgbaimage2[:, :, 3] = alpharamp[::-1]
 
-p1 = Picture(rgbaimage1, channels=4)
+p1 = Image(rgbaimage1, channels=4)
 
-p2 = Picture(rgbaimage2, channels=4).z(12)
+p2 = Image(rgbaimage2, channels=4).z(12)
 
 show(p1, p2, __doc__, axes=7, viewup="z").close()
 
@@ -24,6 +24,6 @@ img[256:,  0:256] = 128
 img[256:,   256:] = 255
 img = img.transpose(1,0)
 
-pict = Picture(img)
+pict = Image(img)
 
 show(pict, mode="image", bg=(0.4,0.5,0.6), axes=1).close()